+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Improvements:
  * icpp: allow io cache object resolve imports lazily.
  * icpp: move all the boost wrapper APIs to icppex header to improve portability;
  * icpp: generate io cache file anyway even in temporary directory;
+ * idbg: add the support of icppdbg library for Cutter++;
 
 ## v0.1.3
 Bug fixes: